fix(bootstrap_microk8s.sh): fix shebang line and add check for existing microk8s installation
The shebang line was fixed to properly indicate the shell used. A check
was added to verify if microk8s is already installed before attempting
to install it. The user is now added to the microk8s group to allow
running microk8s commands without sudo. DNS is enabled to allow for
service discovery. Finally, the kubeconfig is generated to allow kubectl
to connect to the microk8s cluster.

+++ scripts/bootstrap_microk8s.sh
@@ -1,2 +1,8 @@
-!#/bin/sh
-sudo snap install microk8s --classic --channel=1.26/stable
+if ! [ -f `which microk8s` ];
+then
+    sudo snap install microk8s --classic --channel=1.26/stable
+fi
+sudo usermod -a -G microk8s $USER
+sudo microk8s status --wait-ready
+sudo microk8s enable dns
+sudo microk8s.config > kubeconfig
